Fixed errors preventing compiling in ByteArray and its dependencies. Initial unit tests were also written. Modified the bit operations to use uintmax_t instead of size_t for the bit index, since we can have a byte array of size SIZE_MAX, and we would need 8*SIZE_MAX bit indices, a number trivially greater than SIZE_MAX. Renamed String_example.h to something more indicative of its content. Finally wrote down .ps1 compilation command. Still need to add to profile and codesign.

+++ src/countZeroBits.c
@@ -1,7 +1,7 @@
 #include "countZeroBits.h"
 
-size_t countZeroBits(size_t size, unsigned char* bytes){
-  return size*8 - countOneBits(size, bytes);
+uintmax_t countZeroBits(size_t size, unsigned char* bytes){
+  return (uintmax_t)size * 8 - countOneBits(size, bytes);
 }
 
 #ifdef TEST_COUNT_ZERO_BITS
